+++ src/print.rs
@@ -14,6 +14,14 @@ pub fn print_system_time_to_rfc3339(time: &SystemTime) -> String {
 }
 
 
+/// Print a `SystemTime` as a RFC3339 time stamp.
+pub fn print_system_time_to_rfc3339_with_nanos(time: &SystemTime) -> String {
+  // Rust's `SystemTime` internally work with nano seconds and so by
+  // doing the same we hope to have no loss of information.
+  DateTime::<Utc>::from(*time).to_rfc3339_opts(SecondsFormat::Nanos, true)
+}
+
+
 #[cfg(test)]
 mod tests {
   use super::*;
@@ -27,6 +35,17 @@ mod tests {
     let string = "2018-04-01T12:04:17.050Z";
     let time = parse_system_time_from_str(string).unwrap();
     let result = print_system_time_to_rfc3339(&time);
+    assert_eq!(result, string)
+  }
+
+
+  /// Check that we can format a `SystemTime` as a RFC3339 time stamp
+  /// with nanosecond resolution.
+  #[test]
+  fn print_rfc3339_with_nanos() {
+    let string = "2018-04-01T12:04:17.050937231Z";
+    let time = parse_system_time_from_str(string).unwrap();
+    let result = print_system_time_to_rfc3339_with_nanos(&time);
     assert_eq!(result, string)
   }
 }